New component structure
fix https://github.com/Kamahl19/react-starter/issues/38

fix https://github.com/Kamahl19/react-starter/issues/17
