+++ package.json
@@ -45,7 +45,8 @@
     "react-scripts": "^2.1.3",
     "redux-immutable-state-invariant": "^2.1.0",
     "redux-logger": "^3.0.6",
-    "source-map-explorer": "^1.6.0"
+    "source-map-explorer": "^1.6.0",
+    "storybook-react-router": "^1.0.2"
   },
   "dependencies": {
     "antd": "^3.13.0",
