--- src/features/auth/components/AuthSpinner/style.less
@@ -1,10 +1,0 @@
-.auth-spinner {
-  text-align: center;
-  .anticon {
-    @size: 48px;
-    font-size: @size;
-    width: @size;
-    height: @size;
-    margin-bottom: @padding-lg;
-  }
-}
