+++ src/features/auth/components/ForgottenPassword/stories.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import ForgottenPassword from './component';
+
+storiesOf('ForgottenPassword', module).add('default', () => <ForgottenPassword />);
