+++ src/features/auth/components/Login/stories.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import Login from './component';
+
+storiesOf('Login', module).add('default', () => <Login />);
