+++ src/features/auth/components/ResetPassword/index.js
@@ -1,47 +1,45 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Trans, withTranslation } from 'react-i18next';
+import { connect } from 'react-redux';
 
-import { FormScreen, FormItem } from '../../../../packages/ant-form-helpers';
+import { connectSpinner } from '../../../../packages/spinner';
+import { resetPasswordRequest } from '../../ducks';
+import { apiCallIds } from '../../api';
 
-import { Button, Form, Input } from '../../../../common/components';
-import rules from '../../../../common/rules';
+import ResetPassword from './component';
 
-const ResetPasswordForm = ({ form, isLoading, t, onSubmit }) => (
-  <FormScreen form={form} onSubmit={onSubmit}>
-    {({ hasErrors, handleSubmit }) => (
-      <Form onSubmit={handleSubmit}>
-        <FormItem
-          id="email"
-          rules={[rules.required, rules.email]}
-          label={<Trans i18nKey="fields.email.label">E-mail</Trans>}
-        >
-          <Input placeholder={t('fields.email.placeholder', { defaultValue: 'E-mail' })} />
-        </FormItem>
-        <FormItem
-          id="password"
-          rules={rules.passwordWithLimit}
-          label={<Trans i18nKey="resetPassword.newPassword.label">New Password</Trans>}
-        >
-          <Input.Password
-            placeholder={t('resetPassword.password.placeholder', {
-              defaultValue: 'Enter New Password',
-            })}
-          />
-        </FormItem>
-        <Button block type="primary" htmlType="submit" loading={isLoading} disabled={hasErrors}>
-          <Trans i18nKey="fields.submit">Submit</Trans>
-        </Button>
-      </Form>
-    )}
-  </FormScreen>
-);
-
-ResetPasswordForm.propTypes = {
-  form: PropTypes.object.isRequired,
-  isLoading: PropTypes.bool.isRequired,
-  t: PropTypes.func.isRequired,
-  onSubmit: PropTypes.func.isRequired,
+const mapDispatchToProps = {
+  resetPassword: resetPasswordRequest,
 };
 
-export default withTranslation()(Form.create()(ResetPasswordForm));
+const ResetPasswordContainer = ({
+  isLoading,
+  match: {
+    params: { passwordResetToken },
+  },
+  resetPassword,
+}) => (
+  <ResetPassword
+    isLoading={isLoading}
+    onSubmit={data => resetPassword({ ...data, passwordResetToken })}
+  />
+);
+
+ResetPasswordContainer.propTypes = {
+  isLoading: PropTypes.bool.isRequired,
+  match: PropTypes.shape({
+    params: PropTypes.shape({
+      passwordResetToken: PropTypes.string.isRequired,
+    }),
+  }),
+  resetPassword: PropTypes.func.isRequired,
+};
+
+export default connect(
+  undefined,
+  mapDispatchToProps
+)(
+  connectSpinner({
+    isLoading: apiCallIds.RESET_PASSWORD,
+  })(ResetPasswordContainer)
+);
