+++ src/features/auth/components/ResetPassword/stories.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import ResetPassword from './component';
+
+storiesOf('ResetPassword', module).add('default', () => <ResetPassword />);
