+++ src/features/auth/components/SignUp/stories.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import SignUp from './component';
+
+storiesOf('SignUp', module).add('default', () => <SignUp />);
