--- src/features/auth/style.less
@@ -1,1 +1,0 @@
-@import './components/AuthSpinner/style.less';
