+++ storybook/addons/SuspenseDecorator.js
@@ -1,0 +1,3 @@
+import React, { Suspense } from 'react';
+
+export default storyFn => <Suspense fallback={<div>loading</div>}>{storyFn()}</Suspense>;
