+++ storybook/addons/index.js
@@ -1,3 +1,2 @@
 export { default as LocaleDecorator } from './LocaleDecorator';
 export { default as ReduxDecorator } from './ReduxDecorator';
-export { default as Row } from './Row';
